todo_api/requirements.txt
Django==3.2.5
django-cors-headers==3.7.0
django-extensions==3.1.3
djangorestframework==3.12.4
drf-yasg==1.20.0
